Add files via upload
The file login.html and login.css is the first thing the user will see when they open our site. The log-in page is used to validate the user and see if they have a account with us. 
They will be asked for a username and a password.
There will be a submit and a reset button.
If the username and password match our records, they will be redirected to the main page
The sign-up form sends the user to a sign-up page where they will be asked to give their username, password, email, and phone number.
Javascript was also implemented to show an alert after an unsuccessful login attempt.

When the user clicks on the example recipe drop down menu, they will be greeted with categories of what food recipes that they may choose from.
When the user clicks on my recipes, they will see two choices either add recipes or view recipes.
Add recipes is where the user goes to a page where they are asked to give the title of the recipes, description/ingredients, and steps.
View recipes is where you go look at what you submitted.
When the user clicks on my tips and tricks, they will see either add tips and tricks or view tips and tricks.
Add tips and tricks is where the user goes to a page where they are asked to enter their username in and for a text input.
View tips and tricks is where you go look at what you submitted.
When the user clicks on the pictures that are in the creator's favorite they will be sent to the recipe of that image.